Remove reference to deprecated image
The #20553 changed the deprecated hello-node image to echoserver image. However, there is a reference to the old image later in the text. Point 5 implies that there will be a 'Hello World' message. This probably was true for the hello-node image but the echoserver does not print this (it echoes back the request with all of its params). This may be misleading, so change the text to a generic 'app's response'.